+++ src/services/api/index.tsx
@@ -1,1 +1,2 @@
 export { axiosProductInstance, axiosUserInstance } from './api';
+export { CustomError } from './api.type';
